requirements.txt
click==6.7
Flask==1.0.2
Flask-Login==0.4.1
Flask-Session==0.3.1
Flask-SocketIO==3.0.1
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
paho-mqtt==1.3.1
python-engineio==2.2.0
python-socketio==2.0.0
PyYAML==3.13
six==1.11.0
Werkzeug==0.14.1
